wpa_supplicant: Add traffic monitor module
Collects information about ongoing traffic on all active interfaces.
This is done periodically by polling for each interface received and
sent bytes and calculating the delta from the last poll.
Interfaces that disconnect between polls are disregarded on the next
poll as this should be an indication for future traffic.

Change-Id: I49e8c884a2f989859a520b6600639a4ad05c203b
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/20301
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: 4307dd88681f257e71ce41394175231abe227446