P2P: Use another interface operating channel as listen channel
Performing a P2P Device flow such as p2p_listen or
p2p_find, can degrade the performance of an active interface
connection, if the listen frequency is different than the
frequency used by that interface.

To reduce the effect of P2P Device flows on other interfaces,
try changing the listen channel of the P2P Device to match the
operating channel of one of the other active interfaces. This change
will be possible only in case that the listen channel is not forced
externally, and will be delayed to a point where the P2P Device
state machine is idle.

Change-Id: I76f56ac0778eaf942e49d5560469c6a55924972e
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/22674
Tested-by: IWL Jenkins
Reviewed-by: David Spinadel <david.spinadel@intel.com>

x-iwlwifi-hostap-dev-build: f384c3d0c9b375a8037db9e2d370d56cd427e809